requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
pyzbar==0.1.9
requests==2.31.0
groq==0.4.1
numpy==1.26.2
opencv-python==4.8.1.78
huggingface-hub==0.20.1
transformers==4.36.0
torch==2.1.0
